requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
pymysql>=1.1
pydantic-settings>=2.0
python-multipart>=0.0.9
python-dotenv>=1.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
google-generativeai>=0.8
sentence-transformers>=3.0
numpy>=1.26
PyPDF2>=3.0
python-docx>=1.1
